templates/requirements.txt
django==4.1.7
django-extensions==3.2.1
gunicorn==20.1.0
python-dotenv==0.21.1
